requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-telegram-bot==20.7
pandas==2.3.1
scikit-learn==1.7.1
numpy==2.2.5
joblib==1.5.1
matplotlib==3.10.3
seaborn==0.13.2
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.1.1
